GBridge

+++ daly/src/main/java/io/freedriver/daly/bms/checksum/CRCSum.java
@@ -6,5 +6,5 @@ import io.freedriver.daly.bms.checksum.debug.CRC8Debugger;
  * Interface to caclulate a checksum
  */
 public interface CRCSum {
-    int calc(byte[] withCRC, int length, CRC8Debugger debugger);
+    byte calc(byte[] withCRC, int length, CRC8Debugger debugger);
 }
